Don't specify a default instance size, create tiny and micro
In order to let grenade run tempest tests in parallel on virtual
environments we need to create the same tiny and micro guests that
we create in a normal devstack run. Otherwise we blow memory on
the environment, causing a scheduler fail, and sadness. Also no
cake.

Removing the explicit setting of DEFAULT_INSTANCE_TYPE will make
tempest create appropriate instance sizes, and use them in it's
guests.

Change-Id: I6adde22499853d0cde53771e83ebf5c5f7041c11
Partial-Bug: 1252170