Remove stop nova processes from upgrade-nova
stop-base already stop services, so don't need to do it again in
upgrade-nova. This is needed So we can keep the old nova-compute up for
rolling upgrade testing. Instead of killing the entire screen process
kill all the services inside of it.

stop-base is the preferred place to manage the stop logic, because it
has access to the base devstack's stop-* commands.

Change-Id: Ifa2b6305f1f6bbe41407ca12a0f3cbc99b09f301