Add rolling upgrade test for nova-compute
Add rolling upgrade test for nova-compute. Nova wants to support the ability
to run stable/havana nova-compute with icehouse everything else. This sets
up a test for that.

When $DO_NOT_UPGRADE_SERVICES contains "n-cpu", the final environment
will be trunk everything except havana nova-compute.  Although
$DO_NOT_UPGRADE_SERVICES takes a list of services other parts of grenade
have to be fixed to support values besides "n-cpu", but a list is used
because in the future we want to support adding other values to this
list.

Change-Id: Ifb733ee2fe0c5bba61b43ae19dca7040c12f7893