Skip exercises that use swiftclient cli on grizzly
With the push of v2.0 swiftclient the install of swfitclient on grizzly
stops working because the version is capped <2. This causes pip to
freak out replace the bin/swift script but leave the code from master
at >2 in the swiftclient namespace. This disconnect causes the exercises
that use the swiftclient cli to fail with import errors in the bin/swift
script. This commit skips these exercises to avoid the failures.

Change-Id: Icfc1bd1a6e846bddf63840f09e45bd74920fa162