add a more comprehensive check-sanity
this ensures that all the services we believe are enabled in the
new environment are actually started after we have come up on the
new side. It does couple us to the devstack service list in a way
that we weren't before, however that's probably a long term useful
place to be.

This should make the discovery of failed started services a lot
more reliable.

Remove old `service_check` as that only failed if there was a
failure file, and not if the service never started up in bash.

Change-Id: I4fbafbcb8ac5c16ec79ded231e9a05b0ccbc04fc
(cherry picked from commit dcfde08f2eea0281e021bf8c8f3edd831c364199)