provide timestamps in grenade output
this is part of the effort to make grenade summaries actually work
in the gate like we expect, so that the base case can take the
output out of the main console and into summary consoles for
readability.

Instead of duplicating the awk filter from devstack, create a new
python tool for this, which lets us have much greater control and
sanity in filtering out garbage that we don't care about.

Change-Id: I235a725b6ca8a02caddad9f5a1457f6461269d5b