Add within-juno dir and within-master symlink
This adds logic to upgrade_project() to call project upgrade
scripts from the within-$release directory if performing a
sideways upgrade within the same release.  A symlink named
within-master is added and linked to the current development
release to allow sideways migration testing of trunk.  When Juno
tags are cut and K opens, a from-kilo directoy should be created
and this symlink updated.

Change-Id: I770075442d05f38515e5dadf24f2c642d14d7845