Fix check-sanity + save-state /w disabled services
This adds a mapping for all process names described in BASE_SERVICES
and TARGET_SERVICES, allowing check-sanity to skip checking for running
processes of services that are disabled according to ENABLED_SERVICES.
It also updates grenade.sh and save-state to only dump databases that
exist.  These changes allow grenade to pass against deployments that
have disabled non-essential services in devstack.

Change-Id: Ic4b2bab88b12f33f939b9bf7aca47e930d799f3f
Closes-bug: #1355499