Only run worldump.py if it is there
Check for worldump being present and executable before trying to run
it. Otherwise an error can happen that masks the error that caused
an early exit.

worlddump.py will not be present if the target devstack directory
is not yet there (has not been git cloned yet). Generally speaking
this will only happen when making (and testing) changes to grenade
itself or when something has gone very wrong quite early in a grenade
run.

The target release directory is used because at the current time
worldump.py is present in Juno but not in Icehouse.

Change-Id: I974f5bea23464b4143db8b2d35927d5350921bd2