Don't stop screens in upgrade-*
Stopping things is the job of stop-base.

upgrade-horizon is somewhat special in this case because stop-base
does nothing with it, so just to be sure it continues to stop but
uses screen_stop instead.

In the process remove direct references to calling 'screen' which
allows USE_SCREEN=False to work without 'screen' being unable to
find screens.

Change-Id: I7cf36feee0006c86f39c00b695adf049b87a1251