Remove commented out codes and files related to upgrade-packages
This commit removes some unnecessary commented out codes and files they
are related to upgrade-packages. Upgrading package was part of the
original vision of Grenade. However the scope of upgrades was cut back
now. Because any package upgrades required should be handled in the
service that requires them as Dean Troyer said.

Change-Id: Ic1b6ee2e2b668c0c76c2a1425475651bf8a8bd73