Also restart nova-api as part of nova-bm -> ironic migration
The nova-api service needs a restart after the migration to pick up the new
compute_driver config setting.  This toggles whether baremetal node requests
are proxied to ironic or served locally from the nova-bm database.  A new
test was added to tempest that stresses this and is failing the juno grenade
job because the API is attempting to serve nodes locally instead of via
Ironic.

Change-Id: I4f8354abfd909ba4acd62d3ce4a6ab0535a64f7c
Closes-bug: #1422520