fix juno upgrade-keystone script
The following fixes the upgrade-keystone script for juno. There was a
copy/paste error in the original version of this script, which we
landed, which sets the version discovery of admin urls to public
urls. This means that 'openstack user list' will not function after
this upgrade.

Nothing currently tests this in our stack, which is how this got
through. The attempt to add keystone resource scripts in module
grenade exposed it.

Change-Id: I08d0a2992d0cf28756fb7d08aef6b67b0e597748