add nova modular resource checks
This adds nova modular service checks, including the use of the
verify_noapi phase to ensure the server doesn't go away when
nova-compute is down (which is a substantial enhancement of our
existing validation).

There is slight regression in checking multiple servers, however that
did not seem to be a bug that we actually ran into.

add our own local copy of ping_check to grenade. The version in
devstack gets diverted down a neutron path even though we actually
want to use it in a public context. We need to change that code to
work on public networks regardless of backend. Until then, we'll keep
our own copy.

Change-Id: I985ca38ef670b03802f9ad67ce19cf94cfc802ae