fix neutron grenade due to kilo rename of lib/neutron
Now that we are upgrading from kilo -> liberty, we need to reference
the correct filename in the shutdown portion.

Also fix import order so that we pick up TARGET devstack dir
early. There is I think *a lot* of this file that can actually be
deleted, but that's for another day.

Depends-On: I02a7af995dc7de857c4efcf2cef2f95d357c007a

Change-Id: Ib14321fc59969e720db289c44ca297dd88a94146