Nova's flavor migration should be run within-kilo not from-kilo
We actually need to run migrate_flavor_data on kilo code, before we upgrade to
the new code. We want to remove that compatibility code, and this started to
fail when we tried. So, this patch makes sure that we run the within-kilo step
(as provided for in the plugin docs) and do the migrate_flavor_data there.

Change-Id: I9daa648f803218bf1d8b7b7686ab73e146eeb04b