Wait for nova_server1 to be deleted before deleting it's security group
This adds a 30 second wait loop for the nova_server1 resource to be
deleted before trying to delete it's related security group so we don't
hit a 400 error from nova-api.

This is basically a copy of what is done when deleting the
cinder_server1 resource.

Added notes that we can clean this up with the --wait option on the
`openstack server delete` command when python-openstackclient>=1.4.0 at
least in global-requirements, which will have to wait until at least the
M release since stable/kilo caps osc at <1.1.0.

Closes-Bug: #1460112

Change-Id: I16981765c12db7145f7c4754d454431afb6d5cbb