Add docs building support to grenade
This commit adds basic docs building support to grenade. Previously
we just had rst files in the root of the repo and relied on github
or something else to render them. After this commit we'll support
sphinx built docs.

The next step after this is to add a docs publishing job to infra so
we'll have hosted grenade documentation like other projects.

Change-Id: I505c1d5e250d103b6e0f9da008f44d3ce28df94f