remove additional service checks for upgrading
Even though you've told grenade to UPGRADE a project, we still go and
check to see if it's enabled in devstack before we believe you. For
components that are fully implemented in the grenade and devstack
trees, this is fine. However for plugins this is weird, because
stackrc doesn't provide enough information to figure this out.

We should not do this second check, just trust your grenade config is
correct. This code was probably vestigial to getting the refactor
working correctly before the kilo release.

Change-Id: I01d00f95d602b9b159b3b64ff46a7cd3e23c3983