export DATABASES_TO_SAVE for save-state
I noticed this in I45894911c0ee9cd6c9b142d6192350269873d82f where the
"old" database isn't dumped.  This seems to be because "save-state"
calls inc/upgrade:save_mysql_dbs, but it doesn't have access to
DATABASES_TO_SAVE because it's not exported.  An additional trick is
that save-state includes inc/upgrade, so be sure not to overwrite the
variable as blank.

Depends-On: I45894911c0ee9cd6c9b142d6192350269873d82f
Change-Id: I49626dcc277497eaf66d4d998ba8b2655293d28f