remove partial upgrade support for nova
Partial upgrade support was a very hacky hybrid state of no stopping
some services so not everything is upgraded to new code levels. Now
that we have a multinode upgrade model in the gate we don't need this
any more, and we should burn this with fire so that more folks don't
go down this path.

Change-Id: I2b1b440aa3cccfb2b72e65e41809c359438024be