Grenade should install prereqs on the new side
When we do an upgrade, there may be new rpms/debs that may be
needed. For example, Swift does not have liberasurecode1 in
stable/liberty as the older pyeclib downloads, compiles and
installs things it needs. However the newer pyeclib does not
and the devstack master has been updated to install newer
items needed by pyeclib in:
Idbc2ca3f677f1b8153ebf3a5912f4354525a55c7

So when we do an upgrade, we should run install_prereqs.sh
from the new devstack to lay down the new dependencies.

Change-Id: Ied844017732461661e3c24e2ce38b413c881bea9