Add cell0 database during Nova upgrade
Nova now has a third database that needs to be initialized. It will
reside on the same host as the nova_api database, and migrations are
handled by the "nova-manage db sync" command.

If we're coming from Mitaka the database needs to be created as that
wasn't done in Mitaka Devstack.

Note that this is technically optional for upgrading to Newton, so
it's conditional here so we can have a mix of CI jobs that run Newton
with this setup and without this setup.

The release note for this in nova was added in:

24f0c5b9d6136fe18c3fba0ddd64dab99f6f1aa5

Change-Id: I682713a7a4389e631a4b7a163d5005a6d90463f8