doc: Add a note about communicating exceptions to other teams
Nova caught a few projects off guard with the fact that cells v2
and the placement service are going to be required in Ocata and
to upgrade to Ocata, deployment tooling needs to be potentially
changed in Newton to allow the upgrade. Within our QA/CI system
we handle things like this using Grenade, but we as developers
sometimes forget to communicate these types of major changes
to the various deployment projects in OpenStack, like TripleO.
Developers have gotten used to assuming that everyone reads and
understands the implications of the release notes, and we need
reminders of how to effectively communicate beyond those. This
patch tries to help fill some of that reminder gap.

Change-Id: Ia654b88309a9d2506766e21309317ad14d47099a