Run the placement service when upgrading to Ocata
Now the placement service is mandatory in Ocata, we need to modify Grenade
to deploy that service even if the source local.conf is not providing it.

Note that we need to do this as part of nova right now because the
code for placement is in the nova tree. If we do it before and separately,
we end up starting a newton placement service, which is not enough for
ocata nova (and/or we partially upgrade nova from 55_placement, which
breaks within-$release and my brain).

Change-Id: I573149b9415da2a8bb3951a4c4ce71c4c3e48c6f