Implement test for volume attach/detach across releases.
This test ensures that changes to support the new cinder v3 api
do not prevent attachements made in previous releases from being
detached.

Specifically, this changes creates a volume and attaches it to a
server in the old release, and then during post-upgrade
verification ensures that the volume can be detached successfully.

Partially Implements: blueprint cinder-new-attach-apis

Change-Id: I880157c72d40024f9a4aa4f288d583d1f1468641