Fix save_mysql_dbs for missing dbs
In change I075eb5a88113acfa36519e2c6e2aab87836be065 to workaround the
potentially missing dbs in a cells environment it turned off exiting on
error. This is a bad idea since it will mask real failures. This commmit
fixes the function to check if the db exists before trying to do a
mysqldump and report whether it's missing or not.

Change-Id: I54aac82e412a298cd1d2066a0fa51131a6da2af0