Handle a stestr results directory
Tempest is in the process of switching the tempest run command to use
stestr internally instead of testrepository. This means that results
will be stored in .stestr instead of .testrepository now. This commit
updates the base job tempest run post-processing to check which
directory exists before attempting to copy the results dir.

Change-Id: Ic1fa3a98b6bcd151c489b078028687892655a19b