Add a venv tox job
The docs CI job is failing on trying to run tox -evenv to run build
sphinx. However this job was not defined anywhere. THis commit adds the
job definition to hopefully unbreak the job. To enable the jobs to work
we also need to install the sphinx deps to build the docs. This was
previously done in the tox.ini but this commit moves it to
test-requirements.txt to ensure that all the tox jobs are able to build
docs.

Change-Id: I5457f3e442fe673a81925a8ed2f36f4eab0639cb