Shutdown placement at the end of the base run
Placement API was not being shutdown at the end of the OLD target,
thus when it was being started at the beginning of the NEW target it
was not actually getting new code. A call to stop_placement is added.

In addition, make sure placement is installed in upgrade.sh so that
any new stuff is in place.

The depends-on is to a devstack change which makes sure that
lib/placement does not call remove_uwsgi_config when stop_placement is
called. Without this, there's no config file for the
devstack@placement-api systemd unit to run, and there is an immediate
exit.

Change-Id: I7f2158aeaef82a47e11c6e29675e542023fff4be
Depends-On: Iee763adf7895145d97b184924896db3f1f48a015
Closes-Bug: #1736385