Save and verify allocation and inventory values for nova/placement
This makes the nova resources module save information about the host
inventory (first node only), as well as the allocation information
about the test instance before the upgrade. Afterwards in the
verify phase, we compare those values to the current ones to make
sure we have not lost or changed anything during the upgrade.

Change-Id: Ifef2acf7dce17d0fa21baac9da6f8403e69136a4