Document when pre-release upgrade scripts are run
Every time I have to figure out where in the sequence
of upgrade events the per-release within-$base and
from-$base scripts run, I have to dig into the doc.

We have a pretty good document already on the sequence
of phases in a grenade run, but the upgrade section is
pretty fuzzy, and with no mention of when the per-release
scripts run relative to when new code is installed, which
is pretty important if you're going to write one of these
scripts.

So this adds some docs related to when those per-release
upgrade scripts happen relative to installing new code
during the upgrade.

Change-Id: Ic79cd10b2df5fa54a13f73cf49c8f73fa93e9381