Drop NOVA_CONFIGURE_CELLSV2 variable
Cells v2 has been required for nova since Ocata
so we can drop the conditional logic on the
NOVA_CONFIGURE_CELLSV2 variable, set by the (now branch-specific)
neutron-grenade job. The only remaining check for
the variable is in the from-mitaka script (which might
be irrelevant at this point given mitaka and newton
are both end of life).

Change-Id: Ic3101ef5f82b3341772a591669ff96bf9ab72ab6