Fix member role usage for postgresql
Change I3da3530aa73a8a1500116bcefdcba7b947d5e05e in devstack
renamed the "Member" role to "member" which is OK for MySQL
which is case insensitive by default, but not postgresql which
is case sensitive by default. This change fixes the role name
usage in grenade and also removes the reference to the invalid
bug against keystone.

Change-Id: I3e24581c5c77e07edc7c867296e066f40acbbc29
Closes-Bug: #1792983