Add upgrade script for separate placement from nova
This adds the main upgrade steps for placement extraction
from nova in stein. A from-rocky script is added which will
run after openstack/placement is cloned/installed by devstack
but before services are started on the stein side of the
grenade run.

If the CI infrastructure (devstack-gate/zuul) does not clone the
openstack/placement repo then we need to do it ourselves.

Note that until devstack is actually new enough to configure
the extracted placement repo from change
https://review.openstack.org/#/c/600162/ this upgrade
script will noop and we'll continue to deploy and use placement
from nova. This is needed since we have a catch-22 dependency
with that devstack change which can't land without this grenade
change.

The script will copy the placement-related tables from the
nova_api database and put them into the placement database,
which is also created by the upgrade script.

Then, it will write out the placement config file along
with the placement database connection so the placement
service can start on the stein (new) side.

Finally, it will write out the uwsgi config for placement,
and disable nova-placement-api apache site in favor of
placement-api.

Co-Authored-By: Chris Dent <cdent@anticdent.org>

Change-Id: Ia0f19debb442be2b3d04eae238a3d7287393b5eb