nova: do not verify standard resource classes when using ironic
Starting with the Stein series, the ironic virt driver no longer exposes
the standard VCPU/MEMORY/DISK resource classes. If for any reason the
inventory for them is created on the Rocky side of grenade, the verification
phase fails after upgrading to Stein. This breaks ironic-inspector CI.

This patch updates the logic to check for a custom resource class.

Change-Id: I265046d1615a34f55836264ea8f6ce72d32391cf