Add grenade-py3 job
This adds a grenade job which uses python3 on both
the old and new side of the upgrade run. This is part
of the python3-first community wide goal for the Stein
release. The exact version of python3 used depends on
the version of the underlying distro used in devstack
on both sides of the upgrade run.

Change-Id: Icc37e24c5910f189d09bcfed6cfbd348f6c24ab2