Re-install python-openstackclient after swift
When running grenade under python3, swift runs under python2.  Swift
includes openstack client in its requirements so when it installs under
Python 2 it clobbers the existing openstack client at /usr/local/bin and
any previously installed plugins are wiped out because their entry
points aren't seen. This was discovered because the osc-placement
commands were not available to nova inventory verification commands
during the nova upgrade, which comes after swift.

To address this, we do a reinstall at the end of swift's upgrade.
Because python-openstack client is in ENABLED_PYTHON3_PACKAGES,
it will install as Python3 when installed standalone.

Note that the issue being fixed here is python3 specific, not
OpenStack release specific.

Closes-Bug: 1805156
Change-Id: If23619885a4766e243b39eae6041efe4a4d0a3af