Create config file before database migrate
In the from-rocky placement extraction script, this patch reverses
the order to migrate the database and to create the placement.conf.

This is because ``[placement_database]/connection`` should be set
before the database migration shell script is executed if we want
to stamp the database version within the shell script.

Needed-By: https://review.openstack.org/621733
Change-Id: I8fe401814f887c8b29f1c6262d8b9e5263a07bcc