Wait for volume to be available before attaching it
On slower test nodes we're seeing failures where the
2nd cinder volume created is not yet available by the
time we try to attach it to a server which fails with
a 400 InvalidVolume error. This fixes the bug by waiting
for the volume to be available before trying to attach it.

Change-Id: I833d79ecc97ddc844bf156ab64477c7c77424f20
Closes-Bug: #1807520
(cherry picked from commit 2f03ec23fd9f2fdafd9e064f04f9904a67a6cefc)