Upgrade placement independent of nova
Until now we haven't needed this because placement was being upgraded
as part of nova and placement was not requiring a db sync to upgrade.

Now it does, so an upgrade.sh is added to 15_placement, based on
keystone's (which is simpler than nova's), and the repeated parts
in nova's upgrade.sh are removed. a settings file is also added

Change-Id: I5354a6b32b73c613018bbe17f0691c69f7839b5f
Needed-By: https://review.opendev.org/669170