Make docs build again
Docs build was broken with:

TypeError: __init__() got an unexpected keyword argument 'handle_io_errors'

Upgrade sphinx requirements to those of a recent decade, and replace
the oslosphinx extension and requirement with openstackdocstheme.

At the same time, split the deps for the docs build out of
test-requirements.txt and put them in docs/requiremenst.txt as the gods
intended.

Change-Id: I6f82b5c37217ca2bdcf85c1ce07293b3d1026370