Add ensure-virtualenv role
grenade jobs on ussuri is run with train as base branch where
devstack is not default to py3 so grenade jobs run on ussuri gate
is on python2 and thus rely on virtualenv (instead of us standardising
on "python3 -m venv" on the master branch where we are all Python 3).

Use the zuul-role to ensure the system package that provides
"/usr/bin/virtualenv" is installed.

Devstack did the same but grenade jobs are not derived from devstack
job so we need to do it separately.
- https://review.opendev.org/#/q/I2b092bd70a9dbc6ef3a673ba45ca6925f9220bce

Change-Id: Ieaf1f516fb4657851dc059152c41e8c318d59354