Adding document on Extended Maintainance testing policy
As communicated in ML thread
- http://lists.openstack.org/pipermail/openstack-discuss/2020-June/015499.html

grenade team will only be responsible for maintaining the 'Maintained' stable
branches only not the EM. We will start making the grenade job on EM and oldest
'Maintained' as n-v if they start failing.

This commits document that agreement so that it can be followed consistently.

Change-Id: I615a6b88d39ab1c91fdf57b3393cbbca66833671