Use lib neutron-legacy on subnodes in multinode ci jobs
Devstack's lib/neutron shouldn't be used now as it isn't ready (still,
and probably never will be, but that's different story). On the main
node in multinode CI jobs for grenade we already used lib/neutron-legacy
which is fine. But on the subnodes it was still configured to use
lib/neutron instead of lib/neutron-legacy.
This patch is changing that so lib/neutron-legacy will be used on all
nodes in the grenade multinode jobs.

Change-Id: I6e581deeb6b9f2fc0e671882650e14bb6576c1cd