cinder: Wait for encrypted volume to be deleted before removing type
Cinder volume deletion is async and can take time to complete in the
c-vol backend *after* c-api has already returned to the caller.

As such we need to wait until encrypted volumes are deleted fully before
attempting to delete the associated encrypted volume type as this
request will fail when the volume is still being deleted by c-vol.

This change adds a simple waiter to ensure the volume is removed before
removing the type.

Change-Id: I466763ae9fc5a7ad13b498d43b0c16802c1b800b
Closes-Bug: #1907157