Configure the Tempest venv upper constraints properly
We use Tempest master for testing the supported stable
branches, so using master upper constraints works fine for
such master or stable branch testing, but when we need to
use old Tempest in the below cases then master upper constraints
do not work for Tempest venv:
- Testing Extended Maintenance branch
- Testing py2.7 jobs until stable/train with in-tree tempest plugins

grenade run smoke test on old devstack and if old devstack does not
use the master constraints for Tempest then we need to take care of
the same while running the tests too otherwise, it will end up recreating
the Tempest venv and flush the already installed tempest plugins
and their deps.

This commit set the compatible upper constraint
according to TEMPEST_VENV_UPPER_CONSTRAINTS variable, which is
what used in devstack while creating the Tempest venv and install
tempest plugins in that venv.

Closes-Bug: 1922597
Change-Id: I66b5d128c3dbe51289410760ea1cbf7f61c55c2f