Add grenade-skip-level job
This adds a grenade job that specifically skips the last release,
to prove that we can do it. Such a job would facilitate a new
goal of a tick/tock release arrangement where slower-moving operators
could stick on the "tick" release and know that they can upgrade to
the next "tick" release directly.

Currently this tests from wallaby to yoga (master) and is added as
non-voting until the TC resolution passes (if it does).

Change-Id: I5a8fcc1097956ebbbe7b5ea0bed0889eeeca1d80