Do not run grenade-skip-level job on older stable branches
grenade-skip-level is added and supposed to be run from Yoga
release. We have added this job in integrated template which
is in branchless tempest side and applicable for stable branches
also.

We can control to run this job on Yoga and SLURP release onwards but
that leave if anyone running it directly without template So controlling
the branches in the job definition will work for everyone.

Change-Id: Ie560b25f83c3f102f408077f5023f07facf6095d