Use default neutron backend in the base grenade job
When devstack was switching to ML2/OVN as default backend,
in the base grenade job it was explicitly configured to use old default
ML2/OVS backend. It was done like that to avoid problems with upgrade
and change backend in the same job.
Now, as ML2/OVN is default backend in Devstack since at least
couple of cycles it's fine to switch grenade jobs to use default
backend.

Depends-On: https://review.opendev.org/c/openstack/ironic/+/866993
Depends-On: https://review.opendev.org/c/openstack/devstack/+/867065

Change-Id: Iede9fe71d81fc86a92122800d5a16f45442dd79e