[ussuri-only] Add ensure-rust to grenade jobs
Grenade jobs started to fail on older branches (ussuri and train) as
pypi cannot fetch pre-built wheel for cryptography anymore and while
building it during the test job run it lacks rust compiler.

This patch adds ensure-rust as a workaround to be able to build the
wheels and make the grenade job functional again.

Change-Id: I43f6cf95da6b22ee3fb460929685e76d162329c6