Set NOVA_VNC_ENABLED: true for subnode in the grenade job
Currently, because we don't run the n-novnc (console proxy) service on
the subnode, devstack won't enable VNC for guests hosted on the
subnode.

We see occasional failures in the gate when VNC tempest tests attempt
to obtain a VNC console for a guest that happened to get scheduled to
the subnode with the following error:

  Details: {'code': 400, 'message': 'Unavailable console type novnc.'}

This sets NOVA_VNC_ENABLED = true for the subnode only, with the
thinking that the enabled n-novnc service is sufficient for the
controller node.

Jobs that derive from the "devstack" job [1] are already setting
NOVA_VNC_ENABLED: true for both the controller node and subnode.

[1] https://opendev.org/openstack/devstack/src/commit/3ffcc89d5dda1290c0a3ad6462dd41edb1dc35b7/.zuul.yaml#L535

Change-Id: I1acbdb605e0e0d9e5b4604fcd21af850b06823b9