Fix grenade-skip-level job to run on respective stable branch
grenade-skip-level job has specific setting of releases it upgrade
and depends on the setting of grenade_from_branch and grenade_to_branch
that job version is supposed to run on specific stable branch.

To control where this job should run, we used the negative regex which
required to be updated in every release otherwise zuul will pick wrong
version of this job. Better is not to use the branches rule and let zuul
pick the right version of job as per stable branches and job definition.

Change-Id: If5df9bae3c62042493ec7c39ff3da7c09e830619