Remove grenade-skip-level job from stable/zed
grenade-skip-level job is supposed to run on SLURP releases
only. We started testing it from stable/yoga as non voting but
considering zed as not SLURP (SLURP model started after zed)
this job does not get run. But stable/zed branch contain the
job definition which is carry forward from the stable/yoga(it test
stable/wallaby to stable/yoga). This causing the issue now as it
started running on stable/2024.1 with wrong setting.

Removing it from this branch where it is not supposed to run and
respective stable branches where this job needs to run have their
own definition with correct setting.

Change-Id: I990d1a68581a8a2dc35153175f97b366b7fb5456