Fix grenade-skip-level job to use unmaintained/2023.1
Projects are transitioning to Unmaintained and their stable/2023.1
branches were deleted and unmaintained/2023.1 were cut instead, so we
have to use that instead.

Since testing unmaintained things could not be as a requirement, the
patch set the job as non-voting on the check queue and removes from the
gate queue to reduce CI resource usage.

Change-Id: I1ba3fb368e523d31156f45d2e2ef28ba22dfdb91