Increase Cinder SSH timeout from 30s to 60s
Seen in a nova-grenade-multinode job failure today:

In grenade/projects/70_cinder/resources.sh:create, a server is booted
from a volume and attaches one non-bootable non-encrypted volume
followed by one non-bootable encrypted volume. After the volumes attach
successfully, there is an attempt to SSH to the server and it fails due
to a timeout if SSH does not connect within 30 seconds.

After a failure, the server console log is dumped and in it the last
message is "Starting dropbear sshd: OK" which makes it seem likely that
sshd was not up and running yet and that's why SSH failed.

We could increase the timeout here to give the server a bit more time
to get sshd running before giving up.

Closes-Bug: #2066072

Change-Id: I823d1b984fee9f66b11dc58889cc40ab5555dd62