Add os-test-images in required-project for grenade job
grenade job is inherited from the multinode job which is
defined in zuul-job. Devstack lib/tempest script
needs the project os-test-images repo to be present, so
we need to add it in required project in grenade job.
Same way we did for devstack base job
- https://github.com/openstack/devstack/blob/03bc214525c7d7f9dfb6cb855025b70053839a72/.zuul.yaml#L513

It worked in the grenade job on 2024.1 (master till now) because the old node was stable/2024.1
which did not require the os-test-images but on the current master the old node is stable/2024.2
where the grenade job needs this repo to be present.

Change-Id: I020e05403cb2c75eff3aa0fba0f975d0688cf92d