Run nova-status upgrade always
nova-status was not there before Ocata due to that we
had condition to check if nova-status exist or not. As
Ocata and older releases are EOL, we should not have that
command exist check. The check can give false result when
nova-status does not exist due to any reason and in that
case greande will not fail.

Change-Id: I465d4159f04775890bfb0da20cd14c5e1de19cc4