Try to avoid hanging DB dump by not locking
The patch Ifcd31d522015815d2559365bcb8fdcc5c2b868c6 added a 2 minutes
timeout for the DB dumping so at least the grenade job fails fast
instead of waiting for the whole job timeout. But we still need to solve
the hang of the neutron DB dump. The dump output is empty when the hang
happens so I guess it is waiting for taking a lock. So this patch adds
--skip-lock-tables to the command to see if it helps.

Closes-Bug: #2116336

Change-Id: I0d0c2c14a6e3e24bc8914a7f629516fe25e19a10
Signed-off-by: Balazs Gibizer <gibi@redhat.com>